safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/dom-storage-open-expected.txt

Tests that Web Inspector gets populated with DOM storages according to security origins found in the page.

SecurityOrigin: http://127.0.0.1:8000
IsLocalStorage: true
SecurityOrigin: http://127.0.0.1:8000
IsLocalStorage: false
